pyproject.toml
[tool.poetry]
name = "chat-pdf"
version = "0.1.0"
description = ""
authors = []
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.1.9"
langchain-openai = "^0.0.8"
python-dotenv = "^1.0.1"
pypdf4 = "^1.27.0"
pdfplumber = "^0.10.4"
chromadb = "^0.4.24"
openai = "^1.13.3"
tiktoken = "^0.6.0"
pysqlite3-binary = "^0.5.2.post3"


[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
